Remove old e-mail variable code
