+++ etc/di.xml
@@ -52,11 +52,6 @@
         <plugin name="mageSetupEmailAdminhtmlTemplateEditForm" type="FireGento\MageSetup\Plugin\Email\Block\Adminhtml\Template\Edit\Form"/>
     </type>
 
-    <!-- Can be removed after M2.2 EOL-->
-    <type name="Magento\Email\Model\Source\Variables">
-        <plugin name="mageSetupEmailSourceVariables" type="FireGento\MageSetup\Plugin\Email\Model\Source\Variables"/>
-    </type>
-
     <!-- Email variables for versions >= M2.3.0-->
     <type name="Magento\Variable\Model\Source\Variables">
         <arguments>
